fix type in commons
