--- bundles/org.smarthomej.common/src/main/java/org/smarthomej/common/transform/NoOpValueTransformation.java
@@ -1,42 +1,0 @@
-/**
- * Copyright (c) 2010-2021 Contributors to the openHAB project
- * Copyright (c) 2021 Contributors to the SmartHome/J project
- *
- * See the NOTICE file(s) distributed with this work for additional
- * information.
- *
- * This program and the accompanying materials are made available under the
- * terms of the Eclipse Public License 2.0 which is available at
- * http://www.eclipse.org/legal/epl-2.0
- *
- * SPDX-License-Identifier: EPL-2.0
- */
-package org.smarthomej.common.transform;
-
-import java.util.Optional;
-
-import org.eclipse.jdt.annotation.NonNullByDefault;
-
-/**
- * The {@link NoOpValueTransformation} implements a no-op (identity) transformation
- *
- * @author Jan N. Klug - Initial contribution
- */
-@NonNullByDefault
-public class NoOpValueTransformation implements ValueTransformation {
-    private static final NoOpValueTransformation NO_OP_VALUE_TRANSFORMATION = new NoOpValueTransformation();
-
-    @Override
-    public Optional<String> apply(String value) {
-        return Optional.of(value);
-    }
-
-    /**
-     * get the static value transformation for identity
-     *
-     * @return
-     */
-    public static ValueTransformation getInstance() {
-        return NO_OP_VALUE_TRANSFORMATION;
-    }
-}
