--- bundles/org.smarthomej.common/src/main/java/org/smarthomej/common/transform/ValueTransformation.java
@@ -1,35 +1,0 @@
-/**
- * Copyright (c) 2010-2021 Contributors to the openHAB project
- * Copyright (c) 2021 Contributors to the SmartHome/J project
- *
- * See the NOTICE file(s) distributed with this work for additional
- * information.
- *
- * This program and the accompanying materials are made available under the
- * terms of the Eclipse Public License 2.0 which is available at
- * http://www.eclipse.org/legal/epl-2.0
- *
- * SPDX-License-Identifier: EPL-2.0
- */
-package org.smarthomej.common.transform;
-
-import java.util.Optional;
-
-import org.eclipse.jdt.annotation.NonNullByDefault;
-
-/**
- * The {@link ValueTransformation} applies a set of transformations to a value
- *
- * @author Jan N. Klug - Initial contribution
- */
-@NonNullByDefault
-public interface ValueTransformation {
-
-    /**
-     * applies the value transformation to a value
-     *
-     * @param value The value
-     * @return Optional of string representing the transformed value (empty if transformation not present or failed)
-     */
-    Optional<String> apply(String value);
-}
