--- bundles/org.smarthomej.common/src/main/java/org/smarthomej/common/transform/ValueTransformationProvider.java
@@ -1,34 +1,0 @@
-/**
- * Copyright (c) 2010-2021 Contributors to the openHAB project
- * Copyright (c) 2021 Contributors to the SmartHome/J project
- *
- * See the NOTICE file(s) distributed with this work for additional
- * information.
- *
- * This program and the accompanying materials are made available under the
- * terms of the Eclipse Public License 2.0 which is available at
- * http://www.eclipse.org/legal/epl-2.0
- *
- * SPDX-License-Identifier: EPL-2.0
- */
-package org.smarthomej.common.transform;
-
-import org.eclipse.jdt.annotation.NonNullByDefault;
-import org.eclipse.jdt.annotation.Nullable;
-
-/**
- * The {@link ValueTransformationProvider} allows to retrieve a transformation service by name
- *
- * @author Jan N. Klug - Initial contribution
- */
-@NonNullByDefault
-public interface ValueTransformationProvider {
-
-    /**
-     *
-     * @param pattern A transformation pattern, starting with the transformation service
-     *            name, followed by a colon and the transformation itself.
-     * @return
-     */
-    ValueTransformation getValueTransformation(@Nullable String pattern);
-}
