+++ bundles/org.smarthomej.commons/NOTICE
