+++ bundles/org.smarthomej.commons/pom.xml
